chore : strategy package move to common package
