+++ bm-controller/src/main/java/org/benchmarker/bmcontroller/common/beans/RandomNumberConfig.java
@@ -1,7 +1,7 @@
 package org.benchmarker.bmcontroller.common.beans;
 
-import org.benchmarker.bmcontroller.mail.strategy.IRandomCodeGenerator;
-import org.benchmarker.bmcontroller.mail.strategy.RandomCodeGenerator;
+import org.benchmarker.bmcontroller.mail.common.strategy.IRandomCodeGenerator;
+import org.benchmarker.bmcontroller.mail.common.strategy.RandomCodeGenerator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
