+++ bm-controller/src/main/java/org/benchmarker/bmcontroller/mail/service/impl/MailSenderImpl.java
@@ -5,11 +5,10 @@ import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.benchmarker.bmcontroller.mail.common.factory.EmailBodyGenerator;
-import org.benchmarker.bmcontroller.mail.common.factory.EmailVerificationFactory;
 import org.benchmarker.bmcontroller.mail.controller.dto.EmailCertificationDto;
 import org.benchmarker.bmcontroller.mail.controller.dto.EmailResDto;
 import org.benchmarker.bmcontroller.mail.service.IMailSender;
-import org.benchmarker.bmcontroller.mail.strategy.IRandomCodeGenerator;
+import org.benchmarker.bmcontroller.mail.common.strategy.IRandomCodeGenerator;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
